Generalize renaming upfixers
